[LMCache] core: introduce new logger (AMG-53)
We need to be able to save errors and warnings on disk for analysis
during failures. We already save crashes under `/tmp/` but the crash
output itself often isn't enough. Moreover we want to be able to log
output somewhere where performance is bad due to non-fatal events
(e.g. VRAM alloc failures).

The new logger prints to both stdout/stderr AND under
~/.local/state/lmcache/logs. In that directory the current scheme
is: {timestamp}/lmcache-{pid}.log. Here's an example path:
```
~/.local/state/lmcache/logs/2025_10_31_08_35_45/lmcache-1348846.log
```
You can change the above default path to point to a different
directory with the `LMCACHE_NULOG_DIR` env variable. You can also
adjust the logging level with `LMCACHE_NULOG_LEVEL` (the default is
INFO).

Backwards-Compatibility and Log Selection:
The old logger that prints to stdout is not going away yet and it is
still used for most INFO logs. The new logger is used for most error,
warning, and just a couple of INFO logs that are not in the performance
path.